# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Keep this title stable: generated package paths depend on it.
  title: Api
  version: 1.0.0
  description: School administration API
servers:
  - url: /api
    description: Same-origin API
...
paths:
...
  /exam-sessions:
    get:
      operationId: listExamSessions
      tags: [exams]
      summary: List exam sessions
      responses:
        "200":
          description: Exam sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExamSession"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createExamSession
      tags: [exams]
      summary: Create an exam session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExamSessionInput"
      responses:
        "201":
          description: Created exam session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamSession"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /exam-sessions/{id}:
    patch:
      operationId: updateExamSession
      tags: [exams]
      summary: Update an exam session
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExamSessionInput"
      responses:
        "200":
          description: Updated exam session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamSession"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      operationId: deleteExamSession
      tags: [exams]
      summary: Delete a session with no exams
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Session deleted
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /exams:
    get:
      operationId: listExams
      tags: [exams]
      summary: List scheduled exams
      parameters:
        - name: sessionId
          in: query
          schema:
            type: integer
            minimum: 1
        - name: classId
          in: query
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Exam schedule
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exam"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createExam
      tags: [exams]
      summary: Schedule an exam
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExamInput"
      responses:
        "201":
          description: Created exam
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exam"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /exams/{id}:
    patch:
      operationId: updateExam
      tags: [exams]
      summary: Update a scheduled exam
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExamInput"
      responses:
        "200":
          description: Updated exam
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exam"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      operationId: deleteExam
      tags: [exams]
      summary: Delete an exam with no saved grades
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Exam deleted
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /grades:
    get:
      operationId: listGrades
      tags: [grades]
      summary: List saved grades for an exam or class
      parameters:
        - name: studentId
          in: query
          schema:
            type: integer
            minimum: 1
        - name: examId
          in: query
          schema:
            type: integer
            minimum: 1
        - name: classId
          in: query
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Grade records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Grade"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /grades/bulk:
    put:
      operationId: saveGrades
      tags: [grades]
      summary: Save or update grades for one exam
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkGradesInput"
      responses:
        "200":
          description: Saved grade records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Grade"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      summary: List attendance records by date and class
      parameters:
        - name: studentId
          in: query
          schema:
            type: integer
            minimum: 1
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
        - name: classId
          in: query
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Attendance"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: recordAttendance
      tags: [attendance]
      summary: Record or update a student's attendance status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceInput"
      responses:
        "201":
          description: Recorded attendance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendance"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /attendance/bulk:
    post:
      operationId: recordClassAttendance
      tags: [attendance]
      summary: Atomically mark the class's unrecorded students present without overwriting existing statuses
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassAttendanceInput"
      responses:
        "200":
          description: Committed class attendance
          content:
            application/json:
              schema:
                type: object
                required: [savedCount, skippedCount]
                properties:
                  savedCount:
                    type: integer
                  skippedCount:
                    type: integer
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /attendance/{id}:
    delete:
      operationId: deleteAttendance
      tags: [attendance]
      summary: Remove an attendance entry
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Attendance entry removed
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
  responses:
    BadRequest:
      description: The submitted data is invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Unauthorized:
      description: Sign in is required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Forbidden:
      description: This account does not have permission
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    NotFound:
      description: The requested record does not exist
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
    ClassAttendanceInput:
      type: object
      required: [classId, attendanceDate]
      properties:
        classId:
          type: integer
          minimum: 1
        attendanceDate:
          type: string
          format: date
    ApiError:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    AttendanceStatus:
      type: string
      enum: [present, absent, late, excused]
...
    ExamSession:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        academicYear:
          type: string
        active:
          type: boolean
      required: [id, name, academicYear, active]
    ExamSessionInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        academicYear:
          type: string
          minLength: 4
          maxLength: 30
        active:
          type: boolean
      required: [name, academicYear, active]
    Exam:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        sessionName:
          type: string
        classId:
          type: integer
        className:
          type: string
        section:
          type: string
        subjectId:
          type: integer
        subjectName:
          type: string
        examDate:
          type: string
          format: date
        totalMarks:
          type: number
      required: [id, sessionId, sessionName, classId, className, section, subjectId, subjectName, examDate, totalMarks]
    ExamInput:
      type: object
      properties:
        sessionId:
          type: integer
          minimum: 1
        classId:
          type: integer
          minimum: 1
        subjectId:
          type: integer
          minimum: 1
        examDate:
          type: string
          format: date
        totalMarks:
          type: number
          minimum: 1
      required: [sessionId, classId, subjectId, examDate, totalMarks]
    Grade:
      type: object
      properties:
        id:
          type: integer
        examId:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        studentCode:
          type: string
        className:
          type: string
        subjectName:
          type: string
        examTitle:
          type: string
        score:
          type: number
        maxScore:
          type: number
        note:
          type: [string, "null"]
        updatedAt:
          type: string
          format: date-time
      required: [id, examId, studentId, studentName, studentCode, className, subjectName, examTitle, score, maxScore, note, updatedAt]
    GradeInput:
      type: object
      properties:
        studentId:
          type: integer
          minimum: 1
        score:
          type: number
          minimum: 0
        note:
          type: [string, "null"]
          maxLength: 500
      required: [studentId, score]
    BulkGradesInput:
      type: object
      properties:
        examId:
          type: integer
          minimum: 1
        grades:
          type: array
          items:
            $ref: "#/components/schemas/GradeInput"
      required: [examId, grades]
    Attendance:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        studentCode:
          type: string
        barcode:
          type: string
        classId:
          type: integer
        className:
          type: string
        section:
          type: string
        attendanceDate:
          type: string
          format: date
        status:
          $ref: "#/components/schemas/AttendanceStatus"
        note:
          type: [string, "null"]
      required: [id, studentId, studentName, studentCode, barcode, classId, className, section, attendanceDate, status, note]
    AttendanceInput:
      type: object
      properties:
        studentId:
          type: integer
          minimum: 1
        barcode:
          type: string
          minLength: 1
          maxLength: 100
        attendanceDate:
          type: string
          format: date
        status:
          $ref: "#/components/schemas/AttendanceStatus"
        note:
          type: [string, "null"]
          maxLength: 500
      required: [attendanceDate, status]
...