# lib/api-zod/src/generated/types/attendanceStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * School administration API
 * OpenAPI spec version: 1.0.0
 */

export type AttendanceStatus = typeof AttendanceStatus[keyof typeof AttendanceStatus];


export const AttendanceStatus = {
  present: 'present',
  absent: 'absent',
  late: 'late',
  excused: 'excused',
} as const;